requirements.txt
llama-index-vector-stores-chroma
llama-index
sentence-transformers
llama-index-embeddings-huggingface
chromadb
llama-index-llms-groq
fastapi
uvicorn
python-dotenv